+ changed structure of project, added handlers and storage folders
